Static Auto Routing Implementation for Single Directories
The server can now auto-mount any static files in a given _public directory by calling initializeRouter for that directory. This is in preparation for a static-dynamic server framework that allows for a classic folder-structure static site, with select dynamic endpoints able to be added by simply mapping their routes to the functions that serve them.